pyproject.toml
[tool.poetry]
name = "otterai-api"
version = "0.1.0"
description = ""
authors = ["Stuart Leitch <sleitch@silverchair.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
python-docx = "^1.1.2"
requests-toolbelt = "^1.0.0"
python-dotenv = "^1.0.1"
llama-index = "^0.10.65"
openai = "^1.41.0"
llama-index-vector-stores-milvus = "^0.1.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
